api-bloxs/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
bcrypt==4.0.1
Flask-Cors==3.0.10
SQLAlchemy==1.4.45
alembic==1.9.3
cryptography==2.8
PyMySQL==1.0.2
PyJWT==1.7.1